Failsafe als posts niet meer bestaan.

+++ includes/background-process/delete-applications.php
@@ -27,6 +27,9 @@ class SIW_Delete_Applications_Process extends SIW_Background_Process {
 	protected function task( $item ) {
 
 		$order = wc_get_order( $item );
+		if ( false == $order ) {
+			return false;
+		}
 		$order->delete( true );
 
 		return false;
